py/requirements.txt
fuzzywuzzy
GitPython
ipython
matplotlib
MoviePy
pytest
python-Levenshtein
requests
